Integration Backend and Frontend (UNDER WAY)
